==> notebooks/L08_APIs/01_apis.ipynb <==
# APIs and Web Services

**Outcomes**

- Understand the roles of an API
- Be familiar with how to find an API from a service provider's website
- Understand core parts of an API including endpoints, parameters, and authentication

## What is an API?

API stands for *application programming interface*

An API a way for one computer program or service to communicate with and use another program

### API particpants

There are two main participants in an API request, each with a specific role:

1. Producer: this party sets up a service for others to consume. They dictate behaviors and permissions for access and usage
2. Consumer: this party makes a request to the producer's service. They adhere to structure set up by producer

### How are APIs used?

APIs can be used for many purposes:

- Performing actions on a service: e.g. posting to twitter
- Communicating between services inside a company: e.g. running machine learning tasks on user uploaded photos
- Controlled distribution of data: e.g. access to price feeds for a financial asset

In this class we will primarily use an API to request data from a third party provider

## Identifying an API

Many (most?) platforms or companies provide an API for developers to leverage

I typically search for text like "API", "Developers", or "Tool" on a website

Examples:

- [GitHub](https://github.com): "API"
- [Huboi](https://www.huobi.com/en-us/): "API"
- [World Bank](https://data.worldbank.org/): "Tool" -> "API"

## API Mechanics

There are a few core components for interacting with most modern APIs

These are:

- Endpoints: The URLs needed to access a particular part of the API
- Request type: `GET` type for getting data (usually) `POST` for publishing data (usually)
- Query Parameters: part of a URL that follows a `?`
- Payload: Data contained in *body* of request
- Headers: extra data to provide context for request
- Authentication: how to prove your identity and access privledged content

### Endpoints

APIs are often represented as URLs, or web links

Within an API a URL is called an endpoint

A url has the form `scheme://host/path?query`

- `scheme` is usually `http` or `https`
- `host` is something like `github.com`
- `path` specifies which part of an API or website you are accessing
- `query` has additional options or parameters associated with the request

#### Example 

Example url: `https://api.covidcountydata.org/covid_us?location=eq.12045&variable=eq.tests_total`

- Scheme: https
- host: `api.covidcountydata.org`
- path: `covid_us`
- query parameters: `location=eq.12045` and `variable=eq.tests_total`

### Request Type

When one computer sends a message to another computer over http, it is called a request

Requests have a type associated with them

The two most common request types are

1. `GET`: this is used when the consumer would like to **get** data from the provider
2. `POST`: this is used when the consumer would like to **post** or store data with the provider

### Query Parameters

Query parameters come after the path and a `?`

They provide additional details for *what* is being requested

Multiple parameters are separated by `&`

In example for this url: `https://api.covidcountydata.org/covid_us?location=eq.12045&variable=eq.tests_total`, query parameters are:

1. `location`: with value of `eq.12045`
2. `variable`: with value of `eq.tests_total`

###  Payload

A payload is not found in the URL, but is attached to a request

Sometimes referred to as the *body* of the request

A payload is used to attach data to a request

Let's see an example...

#### Example 

Suppose we want to "like" a post on a social media platform

We might make a `POST` request to `https://api.mysocialmedia.com/likes`

The body of the request might then look like:

```json
{
    "user": "spencer_lyon",
    "post_id": 42
}
```

Here `user` is my username on the platform and `42` is the identifier for the post I am liking

### Headers

Headers are additional options passed with network request

They are passed as a key-value pair, like a dictionary in Python

Thier role is to provide context or metadata about the request

They do **not** typically specify data


####  Common Headers

- `Content-Type`: Specifies what type of data is in the Payload. Example `text/plain` or `application/json`
- `APIKey`: passing the api key
- `Authorization`: A different name some services use for passing api key

### Authentication

The most common forms of authentication are:

- OAuth: use a third party provider (like a social media platform) to prove identity
- API keys: API consumer (us) registers directly with provider and is given a key

We'll use API keys in our examples

Sometimes API keys are part of query parameters, other times part of headers
