==> notebooks/L08_APIs/02_labor_statistics.ipynb <==
# Labor Statistics


## Current Population Survey

The [Current Population Survey](https://en.wikipedia.org/wiki/Current_Population_Survey) (CPS) is one of the oldest, largest, and most well-recognized surveys in the United States.

* A version of this survey has been administered since the late 1930s
* The data is used for many major demographic and economic reports
* Considered reliable at the state level and for the 12 largest metropolitan statistical areas
* Administered to ~60,000 households per month with an average response rate of ~90%

### CPS: How is it administered?

One of the strengths of the CPS is it's panel nature:

* When a household is selected for the CPS, they are interviewed for 4 consecutive months
* The household is then interviewed for 4 consecutive months again 8 months later

Consider an example:

Imagine you were selected for the CPS with your first interview starting in February 2020, then you are interviewed in February 2020, March 2020, April 2020, May 2020, February 2021, March 2021, April 2021, and May 2021.

Details on their methodology can be found [online](https://www.census.gov/programs-surveys/cps/technical-documentation/methodology.html)

### CPS: What is asked?

The CPS collects lots of information on demographics and labor status:

* Number of people in the household? Sex? Age? Race? Relationship between each person?
* Working? Full-time or part-time? Hourly pay? Hours worked?
* Looking for work? How? Why?
* Retired?
* Disabilities?

You can find their exact questionaires [online](https://www.census.gov/programs-surveys/cps/technical-documentation/questionnaires.html)

### CPS: Who uses it?

* Government organizations: Bureau of Labor Statistics (BLS) issues a jobs report once a month based on these numbers
* Researchers: Anyone can access the microdata behind the CPS and use it for their own research

## Labor Statistics:

Below are definitions for common terms in labor statistics:

* **Employed**: Individual who did any work at all as a paid employee, worked in their own business or profession, worked on their own farm, worked 15 or more hours as unpaid workers in a family member's business, or were temporarily absent from their jobs or businesses.
  * (Usually) **Part-time**: Individual who usually works less than 35 hours per week (at all jobs combined)
  * (Usually) **Full-time**: Individual who usually works 35 or more hours per week (at all jobs combined)
* **Unemployed**: Individual who is not employed, was available to work, AND made an active effort to find employment sometime during the 4-week period ending with the survey reference week.
* **Labor force**: The sum of employed and unemployed individuals
* **Unemployment rate**: The number of unemployed people as a percentage of the labor force
* **Not in the labor force**: An individual (aged 16 or older) who was neither employed nor unemployed.

Can find more detail in the [BLS documentation](https://www.bls.gov/opub/hom/cps/pdf/cps.pdf)

### Transition Rates

One key statistic that people often analyze are the transition rates between different states of employment:

* *EE*: The fraction of people who were employed during the current month and in the previous month
* *EU*: The fraction of people who were unemployed during the current month and employed in the previous month
* *UE*: The fraction of people who were employed during the current month and unemployed in the previous month
* *UU*: The fraction of people who were unemployed during the current month and in the previous month

Note, `EE + EU + UE + UU = 1` if we hold the labor force constant

### What drives unemployment levels?

**Two possible ways to increase unemployment**

1. A decrease in UE
2. An increase in EU

In the back of your mind, we'd like you to think about which of these channels will have a larger effect on unemployment levels. We will ask you to investigate this formally next class.

### Great Recession vs COVID-19

We will use this class to explore how the changes in the transition rates changed during two periods of economic turbulence in the United States.
